refactor: extracting the index logic from the 'add' and 'rm' commands and refactoring their code
